feat: c:handle_preconnect
A new callback called "handle_preconnect" which is called before the
Mint.HTTP connection is started. This allows you to add headers or
change the state before making the connection.

This is useful for WS servers that use a timestamp signature in a
header, which is rejected if the timestamp is too old. This is run when
reconnected, so it'll be a fresh header.